% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariatePreprocess.R
\name{covariatePreprocess}
\alias{covariatePreprocess}
\title{Preprocess covariate file}
\usage{
covariatePreprocess(
  covariate = NULL,
  maf = NULL,
  output_file = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{covariate}{Covariate file, can be whether an R data frame or the path of a txt file.}

\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{output_file}{Determine whether to export the preprocessed maf as a txt file.}

\item{quiet}{Whether to show notes during processing.}
}
\value{
The preprocessed covariate file.
}
\description{
This function guarantees that the covariate file is available for DriverGenePathway. Two kinds of covariate input
are feasible. It can be whether a txt file or an R data frame. The function returns the preprocessed
covariate. An txt file of the preprocessed covariate is optional to be exported.
}
\examples{
covariate <- system.file("extdata", "gene.covariates.txt", package = "DriverGenePathway")
laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
covariate <- covariatePreprocess(covariate, laml_maf)
}
