% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePreprocess.R
\name{coveragePreprocess}
\alias{coveragePreprocess}
\title{Preprocess coverage file}
\usage{
coveragePreprocess(coverage, maf = NULL, output_file = TRUE, quiet = FALSE)
}
\arguments{
\item{coverage}{coverage file, can be whether an R data frame or the path of a txt file.}

\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{output_file}{Determine whether to export the preprocessed maf as a txt file.}

\item{quiet}{Whether to show notes during processing.}
}
\value{
The preprocessed coverage file.
}
\description{
This function guarantees that the coverage file is available for DriverGenePathway. Two kinds of coverage input
are feasible. It can be whether a txt file or an R data frame. The function returns the preprocessed
coverage. An txt file of the preprocessed coverage is optional to be exported.
}
\examples{
coverage <- system.file("extdata", "coverage.rda", package = "DriverGenePathway")
load(coverage)
laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
coverage <- coveragePreprocess(coverage, laml_maf)
}
