% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverGenes.R
\name{DriverGenes}
\alias{DriverGenes}
\title{Search driver genes}
\usage{
DriverGenes(Mutation = NULL, Coverage = NULL, Covariate = NULL,
  MutationDict = NULL, chr_files_directory = "chr_files_hg19",
  categ_flag = NaN, bmr = 1.2e-06, p_class = "allTest",
  sigThreshold = 0.05, file_stem = "text")
}
\arguments{
\item{Mutation}{Mutation maf data, default to be NULL.}

\item{Coverage}{Coverage raw data, default to be NULL.}

\item{Covariate}{Covariate data which includes expression level, replication time and chromatin
compartment defaultly, and could be altered by users, default to be NULL.}

\item{MutationDict}{Mapping Variant_Classification to mutation effect in mutation data, default to be NULL.}

\item{chr_files_directory}{Chromosome files directory, "hg19" or "hg38".}

\item{categ_flag}{Mutation category number, should be either NaN or numeric, default to be NULL, maximum to be 6.
If set 0 and if the categories in `Mutation` and `Coverage` match, then use the categories directly.
If set from 1 to 6, then create corresponding categories through maximizing mutation entropy.
If set NaN, then 4 categories if output. If the categories in `Mutation` and `Coverage` do not match and the chromosome folder is not available,
then using two given categories, i.e. `missense` and `null+indel`.}

\item{bmr}{The initial background mutation rate is 1.2e-6, and the value alters when
function ends.}

\item{p_class}{Method of hypothesis. "BB" represents Beta Binomial test, "FCPT" represent Fisher combined P-value test,
"LRT" represents Likelihood Ratio test, "CT" represents Convolution test and "PJ" represents 2D-projection method.
The default value is `allTest`, which runs all above hypothesis tests.}

\item{sigThreshold}{The threshhold of q-value to judge if the gene is significant.}

\item{file_stem}{Filestem}
}
\value{
A list of significant genes.
}
\description{
This function defaultly depends on the DriverGenes workspace folder, available to download at \href{http://dplyr.tidyverse.org/articles/programming.html}{DriverGenes_workspace},
which contains a chromosome folder, hg19 or hg38,a coverage txt file and a covariate txt file. Thus in order to run `DriverGenes`, the only necessary input is the MAF
data. `DriverGenes` outputs genes most likely to be driver genes, which includes 3 main sections, preprocessing of input
data, calculation of background mutation rate, and screening genes with a variaty of methods, specificly, Beta Binomial
Test, Fisher’s Combined p-value Test, Likelihood Ratio Test, Convolution Test, and Projection method. The output of `DriverGenes` includes preprocessed results and significant genes,
saved in two folders automaticaly. The preprocessed result is consist of a txt file of mutation categories as well as 2 pdf plots, showing the number of genes of each category and
each effect. The result of significant genes includes top genes with their p-values and q-values saved as csv file. A table of all q_values is also generated.
}
