% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverPathway.R
\name{DriverPathway}
\alias{DriverPathway}
\title{Search driver pathway using de novo method}
\usage{
DriverPathway(mutation_matrix, driver_size = 3, pop_size = 200,
  iters = 500, permut_time = 1000)
}
\arguments{
\item{mutation_matrix}{Matrix of mutations occur in which patient and which gene}

\item{driver_size}{Size of output driver gene set, default to be 3.}

\item{pop_size}{Population size of GA, default to be 200.}

\item{iters}{Time of iteration of GA, default to be 500.}

\item{permut_time}{Time of Permutation test, default to be 1000.}
}
\description{
This function searches driver pathway using de novo method based on mutual exclusivity and coverage
It outputs the driver pathway and its p-value as a txt file.
}
\examples{
data(mutation_matrix)

DriverPathway(mutation_matrix, driver_size=3, pop_size=30, iters=100, permut_time=50)
}
