% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafPreprocess.R
\name{mafPreprocess}
\alias{mafPreprocess}
\title{Preprocess maf}
\usage{
mafPreprocess(maf = NULL, output_file = TRUE, quiet = FALSE)
}
\arguments{
\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{output_file}{Determine whether to export the preprocessed maf as a txt file.}

\item{quiet}{Whether to show notes during processing.}
}
\value{
The preprocessed maf.
}
\description{
This function guarantees that the maf is available for DriverGenePathway. Two kinds of maf input
are feasible. It can be whether a txt file or an R data frame. The function returns the preprocessed
maf. An txt file of the preprocessed maf is optional to be exported.
}
\examples{
laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
maf <- mafPreprocess(laml_maf)
}
