% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafToMatrix.R
\name{mafToMatrix}
\alias{mafToMatrix}
\title{Transform maf to mutation matrix}
\usage{
mafToMatrix(
  maf = NULL,
  threshold = 5,
  analyse_genes = NULL,
  exclude_genes = NULL,
  analyse_patients = NULL,
  exclude_patients = NULL,
  output_file = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{threshold}{Minimum mutation count for genes to be analysed, default to be 5.}

\item{analyse_genes}{Genes to be analysed.}

\item{exclude_genes}{Genes excluded from analysis.}

\item{analyse_patients}{Patrients to be analysed.}

\item{exclude_patients}{Patrients excluded from analysis.}

\item{output_file}{Determine whether to export the preprocessed maf as a txt file.}

\item{quiet}{Whether to show notes during processing.}
}
\description{
This function transforms maf files to mutation matrix in order to search driver pathways.
}
\examples{
laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
mafToMatrix(maf = laml_maf)
}
