% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFunctions.R
\name{preprocessAssignCateg}
\alias{preprocessAssignCateg}
\title{Assign categories for mutation types}
\usage{
preprocessAssignCateg(K)
}
\arguments{
\item{K}{Dataframe of previously detected mutation categories}
}
\description{
For each of the 192 mutation types, assigning a corresponding category from number 1 to 4.
}
