% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocess input data}
\usage{
preprocessing(M, C, dict, V, chr_files_directory = "hg19",
  categ_flag = NaN, output_filestem = "Preprocessed")
}
\arguments{
\item{M}{Mutation maf data}

\item{C}{Coverage raw data}

\item{dict}{Mapping Variant_Classification to mutation effect in mutation data}

\item{V}{Covariate raw data}

\item{chr_files_directory}{Chromosome files directory, hg19 or hg38}

\item{categ_flag}{Mutation category number, should be either NaN or numeric}

\item{output_filestem}{Unified prefix of output data, could be modified as well}
}
\value{
The output is a list includes the preprocessed mutation and coverage data, which are the inputs of BMR function.
}
\description{
This function depends on a chromosome files folder, hg19 or hg38, which includes the txt files of each chromosome.
The hg19 folder is able to download at website (http://www.bioconductor.org/).
This function first guarantees the mutation and coverage data are available.
It generates mutation categories, the number of which is 4 by default and can also be defined by users.
Then all mutations are classified according to the categories.
The output includes a list of preprocessed mutation and coverage data.
The function also generates 4 txt files as intermediate results, i.e. mutation file, coverage file,
categ file and mutcateg_discovery file.
}
