% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverGenes.R
\name{driverGenes}
\alias{driverGenes}
\title{Identify significante genes with hypothetical test methods.}
\usage{
driverGenes(
  bmr = NULL,
  p_class = "all",
  maf = NULL,
  coverage = NULL,
  covariate = NULL,
  original_mutation_rate = 1.2e-06,
  max_neighbors = 50,
  ref_genome = NULL,
  category_num = NULL,
  output_file = TRUE,
  filter = TRUE,
  sigThreshold = 0.1,
  quiet = FALSE
)
}
\arguments{
\item{bmr}{A list output by the backgroundMutationRate function.}

\item{p_class}{Hypothetical test methods. "betaBinomial" represents beta binomial test;
"fisherBinomial" represents Fisher combined P-value test; "likelihoodRatio" represents likelihood
ratio test; "convolution" represents convolution test; "projection" represents 2D projection method;
"all" represents applying all above methods.}

\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{coverage}{Coverage file, can be whether an R data frame or the path of a txt file.}

\item{covariate}{Covariate file, can be whether an R data frame or the path of a txt file.}

\item{original_mutation_rate}{The default background mutation rate is 1.2e-6, which is used to
screen a subset of genes through binomial test.}

\item{max_neighbors}{Max number of gene neighbors used to calculate the background mutation rate
of a specified gene.}

\item{ref_genome}{Reference chromosome, either a folder name or a installed BSgenome package.
Default NULL, tries to auto-detect from installed genomes.}

\item{category_num}{Number of mutation categories, default 4. If the number is 0, categories should exist.}

\item{output_file}{Whether to export the results as csv files and pngs.}

\item{filter}{Whether to filter genes by q-values.}

\item{sigThreshold}{The threshhold of q-value to judge if the gene is significant}

\item{quiet}{Whether to show notes during processing.}
}
\value{
Signifiant genes
}
\description{
This function Applies different hypothetical test methods to identify significant genes. It outputs
the genes and their q-values.
}
\details{
This function searches the significant genes using different hypothetical test methods,
including beta binomial test, Fisher combined P-value test, likelihood ratio test, convolution test and
2D projection method. There are two ways to run driverGenes. Since driverGenes can be treated as
the next step of the backgroundMutationRate function, the output of backgroundMutationRate can
one form of input. The other way is by inputting the maf, coverage and covariate files, and driverGenes
automaticaly runs the backgroundMutationRate procedure.
The output of backgroundMutationRate corresponds to the argument "bmr".
}
\examples{
laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
coverage <- system.file("extdata", "coverage.rda", package = "DriverGenePathway")
covariate <- system.file("extdata", "gene.covariates.txt", package = "DriverGenePathway")
load(coverage)
bmr_result <- backgroundMutationRate(laml_maf, coverage, covariate, original_mutation_rate = 1.2e-6,
                                     max_neighbors = 50, ref_genome = NULL, category_num = 1,
                                     output_file = TRUE, quiet = FALSE)
driverGenes(bmr_result, p_class = "betaBinomial", output_file = TRUE, filter = TRUE,
            sigThreshold = 0.1, quiet = FALSE)

driverGenes(p_class = "betaBinomial", maf = laml_maf, coverage = coverage, covariate = covariate,
original_mutation_rate = 1.2e-6, max_neighbors = 50, ref_genome = NULL, category_num = 1,
output_file = TRUE, filter = TRUE, sigThreshold = 0.1, quiet = FALSE)
}
