% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverPathway.R
\name{driverPathway}
\alias{driverPathway}
\title{Search driver pathway using de novo method}
\usage{
driverPathway(
  mutation_matrix = NULL,
  maf = NULL,
  threshold = 5,
  analyse_genes = NULL,
  exclude_genes = NULL,
  analyse_patients = NULL,
  exclude_patients = NULL,
  background_gene_filter = FALSE,
  coverage = NULL,
  covariate = NULL,
  original_mutation_rate = 1.2e-06,
  ref_genome = NULL,
  category_num = NULL,
  driver_size = 3,
  pop_size = 200,
  iters = 200,
  permut_time = 500,
  output_file = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{mutation_matrix}{Matrix of mutations occur in which patient and which gene.}

\item{maf}{Mutation Annotation Format, can be whether an R data frame or the path of a txt file.}

\item{threshold}{Minimum mutation count for genes to be analysed, default to be 5.}

\item{analyse_genes}{Genes to be analysed.}

\item{exclude_genes}{Genes excluded from analysis.}

\item{analyse_patients}{Patrients to be analysed.}

\item{exclude_patients}{Patrients excluded from analysis.}

\item{background_gene_filter}{Exclude genes having fewer mutations than expected using a binomial test.}

\item{coverage}{Coverage file, can be whether an R data frame or the path of a txt file.}

\item{covariate}{Covariate file, can be whether an R data frame or the path of a txt file.}

\item{original_mutation_rate}{The default background mutation rate is 1.2e-6, which is used to
screen a subset of genes through binomial test.}

\item{ref_genome}{Reference chromosome, either a folder name or a installed BSgenome package.
Default NULL, tries to auto-detect from installed genomes.}

\item{category_num}{Number of mutation categories, default 4. If the number is 0, categories should exist.}

\item{driver_size}{Size of output driver gene set, default to be 3.}

\item{pop_size}{Population size of GA, default to be 200.}

\item{iters}{Time of iteration of GA, default to be 500.}

\item{permut_time}{Time of Permutation test, default to be 1000.}

\item{output_file}{Determine whether to export the preprocessed maf as a txt file.}

\item{quiet}{Whether to show notes during processing.}
}
\description{
This function searches driver pathway using de novo method based on mutual exclusivity and coverage
It outputs the driver pathway and its p-value as a txt file.
}
\examples{
mutation_matrix <- system.file("extdata", "sample_mutation_matrix.rda",
package = "DriverGenePathway")
load(mutation_matrix)
driverPathway(mutation_matrix, driver_size = 3, pop_size = 30, iters = 100, permut_time = 50)

laml_maf <- system.file("extdata", "tcga_laml.maf", package = "DriverGenePathway")
driverPathway(maf = laml_maf, driver_size = 3, pop_size = 30, iters = 100, permut_time = 50)
}
